Issue #720 Aggiunto log per segnalare RPT rifiutata nodo. Corretto errore valorizzazione esito RPT.

+++ jars/orm/src/main/java/it/govpay/bd/pagamento/RptBD.java
@@ -19,6 +19,7 @@
  */
 package it.govpay.bd.pagamento;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Collections;
@@ -304,7 +305,7 @@ public class RptBD extends BasicBD {
 			if(pspRedirectUrl != null)
 				lstUpdateFields.add(new UpdateField(RPT.model().PSP_REDIRECT_URL, pspRedirectUrl));
 			if(esito!= null) 
-				lstUpdateFields.add(new UpdateField(RPT.model().COD_ESITO_PAGAMENTO, esito.getCodifica()));
+				lstUpdateFields.add(new UpdateField(RPT.model().COD_ESITO_PAGAMENTO, BigInteger.valueOf(esito.getCodifica())));
 
 			((JDBCRPTService)this.getRptService()).updateFields(idRpt, lstUpdateFields.toArray(new UpdateField[]{}));
 		} catch (NotImplementedException e) {
